perf(snapshot,tree_builder,buffer, trans, memory): - Changed some u64 type FileNode structures to u32 type to reduce memory usage. BREAKING CHANGE. - Changed the method of tree_builder from recursion to subsequent iteration. - Added fixup and utf16 buffer for Rayon thread. - Added transmission compression processing for snapshot differential data. - Replaced global paths with dynamic paths to avoid memory fragmentation and GC failure. - Fixed the issue of memory target not decreasing during idle time.
chore(README, TOKEN):
- Updated README.
- Changed GITHUB TOKEN.
BREAKING CHANGE.

+++ src-tauri/src/models.rs
@@ -26,14 +26,14 @@ pub struct FileNode {
     pub is_dir: bool,
     #[serde(skip_serializing_if = "Vec::is_empty", default)]
     pub children: Vec<FileNode>,
-    pub file_count: u64,
-    pub dir_count: u64,
+    pub file_count: u32,
+    pub dir_count: u32,
     #[serde(skip_serializing_if = "is_zero", default)]
-    pub modified_time: u64,
+    pub modified_time: u32,
 }
 
 #[inline]
-fn is_zero(n: &u64) -> bool {
+fn is_zero(n: &u32) -> bool {
     *n == 0
 }
